VirtualBox Guest Additions
This is the file for creating guest additions on any VM using VirtualBox. Change the name of the VM from "VM_FILENAME" to the name of the guest machine you want to change the display for. You can also change the screen resolution of the machine to fit the display you need, if you happen to be running on a display with higher or lower resolution. Change the value DESIRED_RESOLUTION. 

*NOTE - You must save this code in a .txt file and save it to the parent folder of your VM disk image for this to work. Also make sure you spell the VM name correctly, if it is misspelled it will not be recognize. Once you have save it and run your VM.

The following Resolutions are supported on VirtualBox:
1280×720  | 1920×1080  |  2560×1440  |  2048×1080  |  3840×2160  |  5120×2880  |  7680×4320
     HD		  FHD	        QHD	         2K		        4K		       5K	              8K