Create Installing MacOS on VirtualBox
These are the extra install codes needed to run a MacOS as a guest OS through VirtualBox. Open command prompt and change the directory with the first line to the file path from, "C:\Program Files\Oracle\VirtualBox\", to your MacOS VM. Once you are in that directory paste the remaining codes into command prompt. 

*Working for BigSur and Catalina